shift test coverage into separate test class
